workflows/category-b/daily-brief-engine/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pg8000>=1.31.0
google-auth>=2.38.0
google-api-python-client>=2.160.0
anthropic>=0.40.0
httpx>=0.27.0
openai>=1.0
google-cloud-secret-manager>=2.21.0
